main.py
#!/usr/bin/env python3
"""
VMware VCP VVF VCF Certificate Management Tool
Enterprise-grade certificate management for VMware infrastructure
"""

import os
import sys
import logging
import argparse
from pathlib import Path
from typing import Dict, Any, Optional

import yaml
import click
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.StreamHandler(sys.stdout),
        logging.FileHandler('vmware-certs.log')
    ]
)

logger = logging.getLogger(__name__)


class VMwareCertManager:
    """VMware Certificate Management Application"""
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """Initialize the certificate manager"""
        self.config = config or self._load_config()
        self.version = "1.0.0"
        logger.info(f"VMware Certificate Manager v{self.version} initialized")
    
    def _load_config(self) -> Dict[str, Any]:
        """Load configuration from file or environment"""
        config_path = Path("config/config.yml")
        
        if config_path.exists():
            with open(config_path, 'r') as f:
                return yaml.safe_load(f)
        
        # Default configuration
        return {
            'app': {
                'name': 'vmware-vcp-vvf-vcf-certs',
                'version': self.version,
                'debug': os.getenv('DEBUG', 'false').lower() == 'true'
            },
            'logging': {
                'level': os.getenv('LOG_LEVEL', 'INFO'),
                'format': 'json'
            },
            'vcenter': {
                'host': os.getenv('VCENTER_HOST', ''),
                'username': os.getenv('VCENTER_USERNAME', ''),
                'password': os.getenv('VCENTER_PASSWORD', ''),
                'port': int(os.getenv('VCENTER_PORT', '443'))
            }
        }
    
    def validate_certificates(self) -> bool:
        """Validate VMware certificates"""
        logger.info("Starting certificate validation")
        
        try:
            # Certificate validation logic would go here
            logger.info("Certificate validation completed successfully")
            return True
        except Exception as e:
            logger.error(f"Certificate validation failed: {e}")
            return False
    
    def renew_certificates(self) -> bool:
        """Renew VMware certificates"""
        logger.info("Starting certificate renewal")
        
        try:
            # Certificate renewal logic would go here
            logger.info("Certificate renewal completed successfully")
            return True
        except Exception as e:
            logger.error(f"Certificate renewal failed: {e}")
            return False
    
    def backup_certificates(self) -> bool:
        """Backup VMware certificates"""
        logger.info("Starting certificate backup")
        
        try:
            # Certificate backup logic would go here
            logger.info("Certificate backup completed successfully")
            return True
        except Exception as e:
            logger.error(f"Certificate backup failed: {e}")
            return False
    
    def run(self) -> int:
        """Run the certificate manager"""
        logger.info("Starting VMware Certificate Manager")
        
        try:
            # Main application logic
            if self.validate_certificates():
                logger.info("All certificates are valid")
            else:
                logger.warning("Some certificates need attention")
            
            return 0
        except Exception as e:
            logger.error(f"Application error: {e}")
            return 1


@click.group()
@click.version_option(version="1.0.0")
@click.option('--debug', is_flag=True, help='Enable debug mode')
@click.option('--config', type=click.Path(exists=True), help='Configuration file path')
def cli(debug: bool, config: Optional[str]):
    """VMware VCP VVF VCF Certificate Management Tool"""
    if debug:
        logging.getLogger().setLevel(logging.DEBUG)
    
    if config:
        os.environ['CONFIG_PATH'] = config


@cli.command()
def validate():
    """Validate VMware certificates"""
    app = VMwareCertManager()
    success = app.validate_certificates()
    sys.exit(0 if success else 1)


@cli.command()
def renew():
    """Renew VMware certificates"""
    app = VMwareCertManager()
    success = app.renew_certificates()
    sys.exit(0 if success else 1)


@cli.command()
def backup():
    """Backup VMware certificates"""
    app = VMwareCertManager()
    success = app.backup_certificates()
    sys.exit(0 if success else 1)


@cli.command()
def run():
    """Run the certificate manager"""
    app = VMwareCertManager()
    sys.exit(app.run())


def main():
    """Main entry point"""
    cli()


if __name__ == "__main__":
    main()